Extract track labels code to new module, improve inline docs
