+++ src/js/annotations/heatmap.js
@@ -1,118 +1,22 @@
+/**
+ * @fileoverview Functions for heatmaps of genome annotations.
+ * Heatmaps provide an easy way to visualize very dense annotation data.
+ * Unlike the rest of Ideogram's graphics, which use SVG, heatmaps are
+ * rendered using the Canvas element.
+ */
+
 import * as d3selection from 'd3-selection';
 
 var d3 = Object.assign({}, d3selection);
 
-/**
- * Starts a timer that, upon expiring, hides the track label.
- *
- * To enable users to copy label content to their clipboard, a timer is
- * used to control when the label disappears.  It starts when the user's
- * cursor leaves the track or the label.  If the user moves the cursor
- * back over the annot or label after the timer starts and before it expires,
- * then the timer is cleared.
- */
-function startHideTrackLabelTimeout(ideo) {
-  if (ideo.config.showTrackLabel === false) return;
-
-  ideo.hideTrackLabelTimeout = window.setTimeout(function () {
-    d3.select('#_ideogramTrackLabel').transition()
-      .duration(500)
-      .style('opacity', 0)
-  }, 250);
-}
+import {
+  startHideTrackLabelTimeout, writeTrackLabelContainer, showTrackLabel
+} from './track-labels'
 
 /**
- * Write label div setup with default styling.
+ * Add canvases that will contain annotations.  One canvas per track.
  */
-function writeTrackLabelContainer(ideo) {
-  d3.select(ideo.config.container + ' #_ideogramTrackLabelContainer')
-    .append('div')
-    .attr('id', '_ideogramTrackLabel')
-    .style('opacity', 0)
-    .style('position', 'absolute')
-    .style('text-align', 'center')
-    .style('padding', '1px')
-    .style('font', '11px sans-serif')
-    .style('background', 'white')
-    .style('line-height', '10px')
-}
-
-function getLabels(ideo) {
-  var annotKeys, reservedWords, labels, heatmaps, i;
-
-  if (ideo.rawAnnots.metadata) {
-    labels = ideo.rawAnnots.metadata.trackLabels;
-    if (ideo.displayedTrackIndexes) {
-      labels = labels.filter(function(d, i) {
-        return ideo.displayedTrackIndexes.includes(i + 1);
-      });
-    }
-  } else if (ideo.config.heatmaps) {
-    labels = [];
-    heatmaps = ideo.config.heatmaps;
-    for (i = 0; i < heatmaps.length; i++) {
-      labels.push(heatmaps[i].key);
-    }
-  } else {
-    annotKeys = ideo.rawAnnots.keys.slice(0);
-    reservedWords = [
-      'name', 'start', 'length', 'trackIndex', 'trackIndexOriginal', 'color'
-    ];
-    labels = annotKeys.filter(d => !reservedWords.includes(d));
-  }
-  labels = labels.join('<br>')
-  // labels = 'foo';
-  // labels = 'foo<br>bar<br>baz<br>moo';
-
-  return labels;
-}
-
-function showTrackLabel(trackCanvas, ideo) {
-    var labels, firstTrackId, firstTrack, trackBox, labelBox, left, ideoBox,
-      top,
-      marginHack = 7; // TODO: Make this dynamic
-
-    clearTimeout(ideo.hideTrackLabelTimeout);
-
-    labels = getLabels(ideo);
-
-    firstTrackId = trackCanvas.id.split('-').slice(0, -1).join('-') + '-0';
-    firstTrack = d3.select('#' + firstTrackId).nodes()[0];
-    trackBox = firstTrack.getBoundingClientRect();
-
-    d3.select('#_ideogramTrackLabel')
-      .interrupt() // Stop any in-progress disapperance
-      .style('top', '')
-      .style('left', '')
-      .style('transform', null)
-      .style('transform', 'rotate(-90deg)')
-      .html(labels);
-
-    labelBox = d3.select('#_ideogramTrackLabel').nodes()[0]
-      .getBoundingClientRect();
-    ideoBox = d3.select(ideo.config.container).nodes()[0]
-      .getBoundingClientRect();
-
-    left = Math.round(trackBox.left + labelBox.width) - trackBox.width - 1;
-    left -= ideoBox.left - marginHack;
-    top = -(labels.split('<br>').length - 2) * trackBox.width + 2;
-
-    d3.select('#_ideogramTrackLabel')
-      .style('opacity', 1) // Make label visible
-      .style('left', left + 'px')
-      .style('top', top + 'px')
-      .style('width', 'max-content')
-      .style('transform-origin', 'bottom left')
-      .style('text-align', 'left')
-      .on('mouseover', function () {
-        clearTimeout(ideo.hideTrackLabelTimeout);
-      })
-      .on('mouseout', function () {
-        startHideTrackLabelTimeout(ideo);
-      });
-  }
-
-function writeCanvases(chr, chrLeft, chrWidth, ideoHeight, ideo) {
+function writeCanvases(chr, chrLeft, ideoHeight, ideo) {
   var j, trackLeft, trackWidth, canvas, context, id,
     contextArray = [],
     numAnnotTracks = ideo.config.numAnnotTracks;
@@ -138,6 +42,9 @@ function writeCanvases(chr, chrLeft, chrWidth, ideoHeight, ideo) {
   return contextArray;
 }
 
+/**
+ * Render annotations on the canvas
+ */
 function fillCanvasAnnots(annots, contextArray, chrWidth, ideoMarginTop) {
   var j, annot, context, x;
 
@@ -152,15 +59,13 @@ function fillCanvasAnnots(annots, contextArray, chrWidth, ideoMarginTop) {
 }
 
 /**
- * Draws a 1D heatmap of annotations along each chromosome.
+ * Draw a 1D heatmap of annotations along each chromosome.
  * Ideal for representing very dense annotation sets in a granular manner
  * without subsampling.
  *
  * TODO:
  * - Support in 'horizontal' orientation
  * - Support after rotating chromosome on click
- *
- * @param annots {Array} Processed annotation objects
  */
 function drawHeatmaps(annotContainers) {
   var annots, chrLeft, contextArray, chrWidth, i, chr,
@@ -180,7 +85,7 @@ function drawHeatmaps(annotContainers) {
     chrWidth = ideo.config.chrWidth;
     chrLeft = ideo._layout.getChromosomeSetYTranslate(i);
 
-    contextArray = writeCanvases(chr, chrLeft, chrWidth, ideoHeight, ideo);
+    contextArray = writeCanvases(chr, chrLeft, ideoHeight, ideo);
     fillCanvasAnnots(annots, contextArray, chrWidth, ideoMarginTop);
   }
 
@@ -193,6 +98,9 @@ function drawHeatmaps(annotContainers) {
   }
 }
 
+/**
+ * Given annotation value (m), should it use the color in this threshold?
+ */
 function shouldUseThresholdColor(m, numThresholds, value, prevThreshold,
   threshold) {
 
@@ -219,6 +127,9 @@ function shouldUseThresholdColor(m, numThresholds, value, prevThreshold,
   );
 }
 
+/**
+ * Determine the color of the heatmap annotation.
+ */
 function getHeatmapAnnotColor(thresholds, value) {
   var m, numThresholds, thresholdList, threshold, tvInt, thresholdColor,
     prevThreshold, useThresholdColor, color;
@@ -245,6 +156,9 @@ function getHeatmapAnnotColor(thresholds, value) {
   return color;
 }
 
+/**
+ * Set color and track index for raw annotation objects.
+ */
 function getNewRawAnnots(heatmapKeyIndexes, rawAnnots, ideo) {
   var j, k, ra, newRa, value, thresholds, color, trackIndex,
     newRas = [];
@@ -291,7 +205,7 @@ function reportPerformance(t0, ideo) {
 }
 
 /**
- * Deserializes compressed annotation data into a format suited for heatmaps.
+ * Deserialize compressed annotation data into a format suited for heatmaps.
  *
  * This enables the annotations to be downloaded from a server without the
  * requested annotations JSON needing to explicitly specify track index or
